# docs/contracts/account-opening-submission.openapi.yaml
openapi: 3.1.0
info:
  title: Indian Bank Account Opening Submission API
  version: 1.0.0
  description: |
    Contract for receiving account opening applications submitted from Adaptive Forms.
servers:
  - url: https://api.indianbank.example.com
    description: Production placeholder
  - url: https://sandbox-api.indianbank.example.com
    description: Sandbox placeholder
paths:
  /v1/account-openings:
    post:
      summary: Submit account opening application
      operationId: submitAccountOpeningApplication
      security:
        - OAuth2ClientCredentials: [account.openings.write]
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountOpeningSubmissionRequest'
            examples:
              default:
                $ref: '#/components/examples/AccountOpeningRequest'
      responses:
        '202':
          description: Submission accepted for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionAcceptedResponse'
              examples:
                accepted:
                  $ref: '#/components/examples/SubmissionAccepted'
        '400':
          description: Validation failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Duplicate submission detected via idempotency key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Business rule validation failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /v1/account-openings/{applicationId}:
    get:
      summary: Fetch application status
      operationId: getAccountOpeningStatus
      security:
        - OAuth2ClientCredentials: [account.openings.read]
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: string
            pattern: '^IB-[0-9]{4}-[A-Z0-9]{8}$'
      responses:
        '200':
          description: Status returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionStatusResponse'
        '404':
          description: Application not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    OAuth2ClientCredentials:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: https://auth.indianbank.example.com/oauth2/token
          scopes:
            account.openings.write: Submit account opening applications
            account.openings.read: Read account opening status

  parameters:
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: true
      schema:
        type: string
        maxLength: 64
      description: Trace identifier propagated across downstream systems.
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        maxLength: 64
      description: Client-provided key to prevent duplicate submissions.

  schemas:
    AccountOpeningSubmissionRequest:
      type: object
      additionalProperties: false
      required:
        - requestContext
        - applicant
        - contact
        - account
        - declaration
      properties:
        requestContext:
          type: object
          additionalProperties: false
          required:
            - sourceSystem
            - sourceChannel
            - submittedAt
          properties:
            sourceSystem:
              type: string
              enum: [AEM-EDGE-FORMS]
            sourceChannel:
              type: string
              enum: [WEB]
            submittedAt:
              type: string
              format: date-time
            formVersion:
              type: string
            journeyId:
              type: string
            branchCode:
              type: string
        applicant:
          type: object
          additionalProperties: false
          required:
            - firstName
            - lastName
            - dateOfBirth
            - panNumber
            - aadhaarNumber
          properties:
            firstName:
              type: string
              minLength: 2
              maxLength: 50
            lastName:
              type: string
              minLength: 1
              maxLength: 50
            dateOfBirth:
              type: string
              format: date
            gender:
              type: string
            maritalStatus:
              type: string
            panNumber:
              type: string
              pattern: '^[A-Z]{5}[0-9]{4}[A-Z]{1}$'
            aadhaarNumber:
              type: string
              pattern: '^[2-9][0-9]{11}$'
        contact:
          type: object
          additionalProperties: false
          required:
            - mobileNumber
            - emailAddress
            - addressLine1
            - city
            - state
            - pinCode
          properties:
            mobileNumber:
              type: string
              pattern: '^[6-9][0-9]{9}$'
            emailAddress:
              type: string
              format: email
            addressLine1:
              type: string
              minLength: 5
              maxLength: 120
            addressLine2:
              type: string
              maxLength: 120
            city:
              type: string
              maxLength: 60
            state:
              type: string
              maxLength: 60
            pinCode:
              type: string
              pattern: '^[1-9][0-9]{5}$'
        account:
          type: object
          additionalProperties: false
          required:
            - accountType
            - openingDeposit
            - debitCardRequired
          properties:
            accountType:
              type: string
              enum: [Savings, Current, Salary]
            openingDeposit:
              type: number
              minimum: 1000
              maximum: 1000000
            debitCardRequired:
              type: string
              enum: [Yes, No]
            additionalServices:
              type: array
              items:
                type: string
                enum: [Net Banking, Mobile Banking, SMS Alerts, Cheque Book]
        documents:
          type: object
          additionalProperties: false
          properties:
            passportPhoto:
              type: object
              additionalProperties: false
              properties:
                fileName:
                  type: string
                contentType:
                  type: string
                checksumSha256:
                  type: string
            addressProof:
              type: object
              additionalProperties: false
              properties:
                fileName:
                  type: string
                contentType:
                  type: string
                checksumSha256:
                  type: string
        declaration:
          type: object
          additionalProperties: false
          required:
            - consentAccepted
          properties:
            consentAccepted:
              type: boolean
              const: true
            consentTimestamp:
              type: string
              format: date-time

    SubmissionAcceptedResponse:
      type: object
      additionalProperties: false
      required:
        - applicationId
        - status
        - receivedAt
      properties:
        applicationId:
          type: string
          pattern: '^IB-[0-9]{4}-[A-Z0-9]{8}$'
        status:
          type: string
          enum: [RECEIVED, IN_REVIEW, NEEDS_ACTION, APPROVED, REJECTED]
        receivedAt:
          type: string
          format: date-time
        nextAction:
          type: string

    SubmissionStatusResponse:
      type: object
      additionalProperties: false
      required:
        - applicationId
        - status
        - lastUpdatedAt
      properties:
        applicationId:
          type: string
        status:
          type: string
          enum: [RECEIVED, IN_REVIEW, NEEDS_ACTION, APPROVED, REJECTED]
        subStatus:
          type: string
        lastUpdatedAt:
          type: string
          format: date-time
        remarks:
          type: string

    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - code
        - message
        - correlationId
      properties:
        code:
          type: string
          examples: [VALIDATION_ERROR]
        message:
          type: string
        correlationId:
          type: string
        details:
          type: array
          items:
            type: object
            additionalProperties: false
            properties:
              field:
                type: string
              issue:
                type: string

  examples:
    AccountOpeningRequest:
      value:
        requestContext:
          sourceSystem: AEM-EDGE-FORMS
          sourceChannel: WEB
          submittedAt: '2026-07-04T11:20:01Z'
          formVersion: '1.0.0'
          journeyId: retail-account-opening
          branchCode: BR0456
        applicant:
          firstName: RAVI
          lastName: KUMAR
          dateOfBirth: '1994-10-12'
          gender: Male
          maritalStatus: Single
          panNumber: ABCDE1234F
          aadhaarNumber: '987654321012'
        contact:
          mobileNumber: '9876543210'
          emailAddress: ravi.kumar@example.com
          addressLine1: 16 MG Road
          addressLine2: Near Metro Station
          city: Chennai
          state: Tamil Nadu
          pinCode: '600001'
        account:
          accountType: Savings
          openingDeposit: 25000
          debitCardRequired: 'Yes'
          additionalServices:
            - Net Banking
            - SMS Alerts
        documents:
          passportPhoto:
            fileName: photo-ravi-kumar.jpg
            contentType: image/jpeg
            checksumSha256: 7f57f00f9a80dbf6f95f0c8686fb4d3f52c4cf4ec9ec3939017f478f9df79e3f
          addressProof:
            fileName: address-proof-ravi-kumar.pdf
            contentType: application/pdf
            checksumSha256: 988f6a4f66df372cd6b22a12f41f86df54513a4fc96cc22ed7f4cbf2c4224ca1
        declaration:
          consentAccepted: true
          consentTimestamp: '2026-07-04T11:19:58Z'
    SubmissionAccepted:
      value:
        applicationId: IB-2026-A7D9X2KQ
        status: RECEIVED
        receivedAt: '2026-07-04T11:20:02Z'
        nextAction: KYC_VERIFICATION_IN_PROGRESS